+++ src/RunicMagic.Controller/Models/EntityRenderingModel.cs
@@ -3,5 +3,5 @@
     [Flags]
     public enum EntityRenderingFlags { None = 0, HasLife = 1, HasAgency = 2, IsTranslucent = 4 }
 
-    public record EntityRenderingModel(int X, int Y, int Width, int Height, string Label, EntityRenderingFlags Flags, bool IsCaster, int? PointingEndX, int? PointingEndY, bool IsIndicateTarget, int? IndicateEndX, int? IndicateEndY);
+    public record EntityRenderingModel(long X, long Y, long Width, long Height, string Label, EntityRenderingFlags Flags, bool IsCaster, long? PointingEndX, long? PointingEndY, bool IsIndicateTarget, long? IndicateEndX, long? IndicateEndY);
 }
