+++ src/RunicMagic.Controller/RuneParsing/SpellParser.cs
@@ -4,13 +4,13 @@ namespace RunicMagic.Controller.RuneParsing
 {
     internal static class SpellParser
     {
-        public static (int parsedTokens, ParsingResult<IExecutableStatement>) Parse(string spellString)
+        public static (long parsedTokens, ParsingResult<IExecutableStatement>) Parse(string spellString)
         {
             var tokenStream = new TokenStream(spellString);
 
             var result = RuneParsingDispatcher.ParseNextRune<IExecutableStatement>(tokenStream);
 
-            return (tokenStream.Index + 1, result);
+            return ((long)(tokenStream.Index + 1), result);
         }
     }
 }
