+++ src/RunicMagic.Controller/Services/WorldRenderingService.cs
@@ -18,14 +18,14 @@ public class WorldRenderingService(WorldModel world, RayCastService rayCast)
         var renderingModels = new List<EntityRenderingModel>();
         foreach (var entity in entities)
         {
-            (int X, int Y)? pointingEnd = null;
+            (long X, long Y)? pointingEnd = null;
             if (entity.PointingDirection.HasValue)
             {
                 var castResult = rayCast.Cast(entity.Id, entity.X, entity.Y, entity.PointingDirection.Value);
                 pointingEnd = (castResult.X, castResult.Y);
             }
 
-            (int X, int Y)? indicateEnd = null;
+            (long X, long Y)? indicateEnd = null;
             if (entity.IndicateTarget?.Direction != null)
             {
                 var direction = entity.IndicateTarget.Direction.Value;
@@ -34,8 +34,8 @@ public class WorldRenderingService(WorldModel world, RayCastService rayCast)
                 var dy = castResult.Y - entity.Y;
                 var dist = Math.Sqrt(dx * dx + dy * dy);
                 var capped = Math.Min(dist, 1000);
-                var endX = (int)Math.Round(entity.X + direction.X * capped);
-                var endY = (int)Math.Round(entity.Y + direction.Y * capped);
+                var endX = (long)Math.Round(entity.X + direction.X * capped);
+                var endY = (long)Math.Round(entity.Y + direction.Y * capped);
                 indicateEnd = (endX, endY);
             }
