+++ src/RunicMagic.Database/Create.sql
@@ -1,5 +1,5 @@
 create table EntityTypes (
-    Id   int          not null constraint PK_EntityTypes primary key,
+    Id   bigint       not null constraint PK_EntityTypes primary key,
     Name nvarchar(50) not null
 );
 
@@ -10,27 +10,27 @@ insert into EntityTypes (Id, Name) values
 
 create table Entities (
     Id            uniqueidentifier not null constraint PK_Entities primary key,
-    EntityTypeId  int              not null constraint FK_Entities_EntityTypes references EntityTypes (Id),
+    EntityTypeId  bigint           not null constraint FK_Entities_EntityTypes references EntityTypes (Id),
     Label         nvarchar(100)    not null,
-    X             int              not null,
-    Y             int              not null,
-    Width         int              not null,
-    Height        int              not null,
+    X             bigint           not null,
+    Y             bigint           not null,
+    Width         bigint           not null,
+    Height        bigint           not null,
     HasAgency     bit              not null constraint DF_Entities_HasAgency default 0,
-    Weight        int              not null,
+    Weight        bigint           not null,
     IsTranslucent bit              not null constraint DF_Entities_IsTranslucent default 0
 );
 
 create table EntityLife (
     EntityId         uniqueidentifier not null constraint PK_EntityLife primary key
                                                constraint FK_EntityLife_Entities references Entities (Id),
-    MaxHitPoints     int              not null,
-    CurrentHitPoints int              not null
+    MaxHitPoints     bigint           not null,
+    CurrentHitPoints bigint           not null
 );
 
 create table EntityCharge (
     EntityId      uniqueidentifier not null constraint PK_EntityCharge primary key
                                             constraint FK_EntityCharge_Entities references Entities (Id),
-    MaxCharge     int              not null,
-    CurrentCharge int              not null
+    MaxCharge     bigint           not null,
+    CurrentCharge bigint           not null
 );
