+++ src/RunicMagic.World/Entity.cs
@@ -9,18 +9,18 @@ public class Entity(EntityId id, EntityType type, string label)
     public EntityId Id { get; } = id;
     public EntityType Type { get; } = type;
     public string Label { get; set; } = label;
-    public int X { get; set; }
-    public int Y { get; set; }
-    public int Width { get; set; }
-    public int Height { get; set; }
+    public long X { get; set; }
+    public long Y { get; set; }
+    public long Width { get; set; }
+    public long Height { get; set; }
 
     public LifeCapability? Life { get; set; }
     public ChargeCapability? Charge { get; set; }
     public bool HasAgency { get; set; }
     public bool IsTranslucent { get; set; }
-    public int Weight { get; set; }
+    public long Weight { get; set; }
     public Func<Entity[]>? Scope { get; set; }
-    public Func<int, ReservoirDraw>? Reservoir { get; set; }
+    public Func<long, ReservoirDraw>? Reservoir { get; set; }
     public Direction? PointingDirection { get; set; }
     public IndicateTarget? IndicateTarget { get; set; }
 }
