+++ src/RunicMagic.World/Execution/Location.cs
@@ -1,3 +1,3 @@
 namespace RunicMagic.World.Execution;
 
-public readonly record struct Location(int X, int Y);
+public readonly record struct Location(long X, long Y);
