+++ src/RunicMagic.World/Execution/Number.cs
@@ -1,3 +1,3 @@
 namespace RunicMagic.World.Execution;
 
-public readonly record struct Number(int Value);
+public readonly record struct Number(long Value);
