+++ src/RunicMagic.World/Geometry/RayCastResult.cs
@@ -1,3 +1,3 @@
 namespace RunicMagic.World.Geometry;
 
-public readonly record struct RayCastResult(Entity? HitEntity, int X, int Y);
+public readonly record struct RayCastResult(Entity? HitEntity, long X, long Y);
