+++ src/RunicMagic.World/TeleportEntityService.cs
@@ -2,7 +2,7 @@ namespace RunicMagic.World;
 
 public class TeleportEntityService
 {
-    public void Teleport(Entity entity, int x, int y)
+    public void Teleport(Entity entity, long x, long y)
     {
         entity.X = x;
         entity.Y = y;
