+++ tests/RunicMagic.Tests/EntityFactoryTests.cs
@@ -14,9 +14,9 @@ public class EntityFactoryTests
 
     private static EntityFactory Factory(WorldModel world) => new(world, Logger);
 
-    private static EntityData CreatureData(int maxHp, int currentHp) => new(
+    private static EntityData CreatureData(long maxHp, long currentHp) => new(
         Id: Guid.NewGuid(),
-        TypeId: (int)EntityType.Creature,
+        TypeId: (long)EntityType.Creature,
         Label: "creature",
         X: 0, Y: 0, Width: 10, Height: 10,
         HasAgency: false,
@@ -24,9 +24,9 @@ public class EntityFactoryTests
         MaxHitPoints: maxHp,
         CurrentHitPoints: currentHp);
 
-    private static EntityData ManaSourceData(int maxCharge, int currentCharge) => new(
+    private static EntityData ManaSourceData(long maxCharge, long currentCharge) => new(
         Id: Guid.NewGuid(),
-        TypeId: (int)EntityType.ManaSource,
+        TypeId: (long)EntityType.ManaSource,
         Label: "mana source",
         X: 0, Y: 0, Width: 10, Height: 10,
         HasAgency: false,
@@ -83,7 +83,7 @@ public class EntityFactoryTests
         var world = new WorldModel();
         var logger = new Mock<ILogger<EntityFactory>>();
         var entity = new EntityFactory(world, logger.Object).Create(new EntityData(
-            Guid.NewGuid(), (int)EntityType.Creature, "lifeless creature",
+            Guid.NewGuid(), (long)EntityType.Creature, "lifeless creature",
             X: 0, Y: 0, Width: 10, Height: 10, HasAgency: false, Weight: 0));
         world.Add(entity);
 
@@ -136,7 +136,7 @@ public class EntityFactoryTests
         var world = new WorldModel();
         var logger = new Mock<ILogger<EntityFactory>>();
         var entity = new EntityFactory(world, logger.Object).Create(new EntityData(
-            Guid.NewGuid(), (int)EntityType.ManaSource, "uncharged source",
+            Guid.NewGuid(), (long)EntityType.ManaSource, "uncharged source",
             X: 0, Y: 0, Width: 10, Height: 10, HasAgency: false, Weight: 0));
         world.Add(entity);
 
@@ -160,7 +160,7 @@ public class EntityFactoryTests
     {
         var world = new WorldModel();
         var caster = Factory(world).Create(CreatureData(maxHp: 100, currentHp: 100) with { X = 0, Y = 0 });
-        var nearby = Factory(world).Create(new EntityData(Guid.NewGuid(), (int)EntityType.Object, "nearby",
+        var nearby = Factory(world).Create(new EntityData(Guid.NewGuid(), (long)EntityType.Object, "nearby",
             X: 10, Y: 0, Width: 10, Height: 10, HasAgency: false, Weight: 0));
         world.Add(caster);
         world.Add(nearby);
@@ -175,7 +175,7 @@ public class EntityFactoryTests
     {
         var world = new WorldModel();
         var caster = Factory(world).Create(CreatureData(maxHp: 100, currentHp: 100) with { X = 0, Y = 0 });
-        var distant = Factory(world).Create(new EntityData(Guid.NewGuid(), (int)EntityType.Object, "distant",
+        var distant = Factory(world).Create(new EntityData(Guid.NewGuid(), (long)EntityType.Object, "distant",
             X: 500, Y: 500, Width: 10, Height: 10, HasAgency: false, Weight: 0));
         world.Add(caster);
         world.Add(distant);
@@ -192,7 +192,7 @@ public class EntityFactoryTests
     {
         var world = new WorldModel();
         var entity = Factory(world).Create(new EntityData(
-            Guid.NewGuid(), (int)EntityType.Object, "rock",
+            Guid.NewGuid(), (long)EntityType.Object, "rock",
             X: 0, Y: 0, Width: 5, Height: 5, HasAgency: false, Weight: 0));
 
         entity.Reservoir.Should().BeNull();
