+++ tests/RunicMagic.Tests/Execution/EntityReferenceRunes/DANTests.cs
@@ -11,7 +11,7 @@ public class DANTests
 {
     private static readonly Direction Right = new(1, 0);
 
-    private static Entity MakeEntity(int x, int y, int width = 100, int height = 100)
+    private static Entity MakeEntity(long x, long y, long width = 100, long height = 100)
     {
         return new Entity(EntityId.New(), EntityType.Object, "test")
         {
